Allow user to override the terminal width
Sometimes we can't detect the width of the terminal.  For example when
piping input to colfmt and sending the output to a pager.  In that
case, the user has to tell us the terminal's width.

For now I'm retaining the error message when terminal width detection
fails.  That will help me detect problems with that detection even
when it's not actually being used.