Allow timestamp columns to be rendered as their age
Sometimes the useful information contained in a timestamp is how much
time has passed since that moment.  By specifying 'age' in a column
spec, that duration is rendered in a succinct format.

I tried to support all popular date formats that are also supported by
Go.  It's easy to add other formats later, if needed.